wrap long lines
This is the first substantive editing by Dan Grayson.  In this commit, we wrap
the long lines, so that future changes will be easier to audit -- the unix
command 'diff' operates on lines, so making the lines shorter will result in
less text being marked as changed as the result of a single minor change.  To
audit this change, just observe that the appearance of the resulting pdf file
does not change.